Enforce strict module hierarchy: no upward imports
Move foundation modules (types, xml-utils, async-mutex, wasm-embed) to
src/tools/lib/, move version.ts to src/cli/, rewrite all imports to be
downward-only. Split monolithic service.test.ts into per-tool test files.
Remove barrel re-export service.ts.

Import rules enforced:
- tools/lib/*: only node: builtins and npm packages
- tools/shared.ts: only ./lib/*
- tools/*.ts: only ./shared.js and ./lib/*
- cli/*.ts: only peers (./)
- server.ts: ./tools/* and ./cli/version.js
- index.ts: ./cli/* and ./server.js
