+++ src/tools/export-cadence-board.test.ts
@@ -1,0 +1,13 @@
+import { describe, it, expect } from "vitest";
+import { exportCadenceBoard } from "./export-cadence-board.js";
+import { isErrorResult } from "./lib/types.js";
+
+describe("exportCadenceBoard", () => {
+  it("returns error on non-Windows platforms", async () => {
+    const result = await exportCadenceBoard("C:/designs/test.brd");
+    expect(isErrorResult(result)).toBe(true);
+    if (isErrorResult(result)) {
+      expect(result.error).toContain("only available on Windows");
+    }
+  });
+});
