+++ src/tools/lib/wasm-embed.ts
