+++ src/tools/lib/xml-utils.test.ts
