+++ src/tools/lib/xml-utils.ts
